refactor: loadServerEnv -> getServerEnv
