+++ lib/base_loader.js
@@ -69,22 +69,50 @@ class EggLoader {
     this.eggPaths = [ this.eggPath ].concat(this.frameworkPaths);
     debug('Loaded eggPaths %j', this.eggPaths);
 
-    /**
-     * 获取当前应用所在的机器环境，统一 serverEnv
-     * ```
-     * serverEnv | 说明
-     * ---       | ---
-     * default   | 默认环境
-     * test      | 交付测试
-     * prod      | 主站生产环境，包括预发，线上服务器
-     * local     | 本地开发环境，就是你的电脑本地启动
-     * unittest  | 单元测试环境，tnpm test, NODE_ENV=test
-     * ```
-     *
-     * @member {String} EggLoader#serverEnv
-     */
-    this.serverEnv = this.loadServerEnv();
+    this.serverEnv = this.getServerEnv();
     debug('Loaded serverEnv %j', this.serverEnv);
+  }
+
+  /**
+   * Get environment of Egg, it's not NODE_ENV
+   *
+   * 1. from `$baseDir/config/serverEnv`
+   * 2. from EGG_SERVER_ENV
+   * 3. from NODE_ENV
+   *
+   * serverEnv | description
+   * ---       | ---
+   * default   | default environment
+   * test      | system integration testing
+   * prod      | production
+   * local     | local on your own computer
+   * unittest  | unit test
+   *
+   * @return {String} serverEnv
+   */
+  getServerEnv() {
+    let serverEnv;
+
+    const envPath = path.join(this.options.baseDir, 'config/serverEnv');
+    if (fs.existsSync(envPath)) {
+      serverEnv = fs.readFileSync(envPath, 'utf8').trim();
+    }
+
+    if (!serverEnv) {
+      serverEnv = process.env.EGG_SERVER_ENV;
+    }
+
+    if (!serverEnv) {
+      if (process.env.NODE_ENV === 'test') {
+        serverEnv = 'unittest';
+      } else if (process.env.NODE_ENV === 'production') {
+        serverEnv = 'default';
+      } else {
+        serverEnv = 'local';
+      }
+    }
+
+    return serverEnv;
   }
 
   /**
@@ -203,36 +231,6 @@ class EggLoader {
     return dirs;
   }
 
-  /**
-   * 获取环境变量
-   *
-   * 1. 从 EGG_SERVER_ENV 获取，一般用于测试
-   * 2. 从 `$baseDir/config/serverEnv` 读取，框架可根据实际情况自行设置
-   * 3. 默认值
-   *
-   * @return {String} serverEnv
-   * @see EggLoader#serverEnv
-   */
-  loadServerEnv() {
-    let serverEnv = process.env.EGG_SERVER_ENV;
-
-    const envPath = path.join(this.options.baseDir, 'config/serverEnv');
-    if (fs.existsSync(envPath)) {
-      serverEnv = fs.readFileSync(envPath, 'utf8').trim();
-    }
-
-    if (!serverEnv) {
-      if (process.env.NODE_ENV === 'test') {
-        serverEnv = 'unittest';
-      } else if (process.env.NODE_ENV === 'production') {
-        serverEnv = 'default';
-      } else {
-        serverEnv = 'local';
-      }
-    }
-
-    return serverEnv;
-  }
 
   /**
    * 获取 {@link EggLoader#frameworkPaths}
